refs #2493
Feature:
Mechanism for raw command implemented and tested.
Note that a fallback was not possible because argparse is catching error "internally" and doesn't raise error when it fails to parse argument.
I decided to implement Solveig first idea to pass a raw ami command: ami command amiCommand [args]
Debug:
During refactoring, we lost the capacity to set the xslt transformation via ami command args. I merged from old pyAMI so that it is now possible again.
  - By default, xslt is text.
  - If set in raw command, when using result.output(), we take transfo from command args
  - If using result.output(xsltblabla), then xsltblabla is taken as transformation (even if output is present in command args)

The lxml transformation package generate object that are not "str". It was causing a problem on some error catching. I fixed it.

git-svn-id: https://lpsc.in2p3.fr/svn/AMI/trunk/AMIWebServiceClient/pyAMI@3503 29852e1a-7735-49d4-a4ac-1f29da7335c7